requirements.txt
Pillow
docxtpl
pandas==2.2.3